test/requirements.txt
browser-use
langchain-openai
python-dotenv
pandas
sqlalchemy
tqdm
playwright
fake-useragent
